Add files via upload
This is a work-experience based milestone project that I completed for the Global Career Accelerator program. The SQLCoding_GameJet PDF are the lines of code I wrote to analyze the data. The GameJet Microtransactions Solution Sheet PDF reveals the reasoning and analysis behind each set of code int he SQLCoding_GameJet PDF.