Add files via upload
This is a work-experience based milestone project that I completed for the Global Career Accelerator program. The SQLCoding_FastKitchen PDF are the lines of code I wrote to analyze the data. The FastKitchenCustomers_SolutionSheet PDF reveals the reasoning and analysis behind each set of code int he SQLCoding_FastKitchen PDF.